requirements.txt
boto3
dateparse
numpy
pandas
pillow
pymupdf
spacy
textractcaller
transformers
trp
